requirements.txt
absl-py==0.5.0
aiodns==1.1.1
aiohttp==3.4.4
aiohttp-socks==0.1.9
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
autopep8==1.4
beautifulsoup4==4.6.3
bleach==3.0.2
boto3==1.7.48
botocore==1.10.48
breadability==0.1.20
Brlapi==0.6.6
bs4==0.0.1
cchardet==2.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloudinary==1.12.0
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.2
django-allauth==0.38.0
django-anymail==3.0
django-appconf==1.0.2
django-bower==5.2.0
django-contrib-comments==1.8.0
django-cors-headers==2.4.0
django-elasticsearch-dsl==0.5.0
django-extensions==2.0.7
django-extra-fields==1.0.0
django-filter==2.0.0
django-hashid-field==2.1.1
django-haystack==2.8.1
django-invitations==1.9.2
django-jet==1.0.7
django-model-utils==3.1.2
django-notifications-hq==1.3
django-organizations==1.1.0
django-request-logging==0.6.2
django-rest-auth==0.9.3
django-rest-invitations==0.1.0
django-rest-knox==3.1.0
django-scheduler==0.8.8
django-storages==1.6.6
django-threadedcomments==1.2
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
drf-haystack==1.8.3
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
future==0.16.0
gast==0.2.0
google-api-python-client==1.4.1
grpcio==1.15.0
gunicorn==19.8.1
h5py==2.8.0
hashids==1.2.0
httplib2==0.11.3
icalendar==4.0.1
idna==2.7
idna-ssl==1.1.0
imageio==2.4.1
IndicoIo==1.2.0
ipaddress==1.0.22
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
Keras==2.2.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.5
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mock==2.0.0
moviepy==0.2.3.5
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.5.6
multidict==4.4.2
mysqlclient==1.3.13
nltk==3.3
numpy==1.15.3
oauth==1.0.1
oauth2client==4.1.2
oauthlib==2.1.0
olefile==0.45.1
pandas==0.23.4
pbr==5.1.0
pdfminer.six==20170720
pexpect==4.2.1
Pillow==5.3.0
pinax-referrals==3.0.3
pkginfo==1.4.2
powerline-status==2.6
protobuf==3.6.1
psutil==5.4.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycairo==1.16.2
pycares==2.3.0
pycodestyle==2.4.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.6.4
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
PyPDF2==1.26.0
pyRFC3339==1.0
pyshorteners==0.6.1
PySocks==1.6.8
pysrt==1.1.1
pytest-runner==4.2
python-apt==1.6.2
python-dateutil==2.7.3
python-debian==0.1.32
python-docx==0.8.7
python-slugify==1.2.5
python3-openid==3.1.0
pytube==9.3.3
pytz==2018.6
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
readme-renderer==23.0
reportlab==3.4.0
requests==2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.1.13
schedule==0.5.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
social-auth-core==1.7.0
sumy==0.7.0
system-service==0.3
systemd-python==234
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
testresources==2.0.1
tqdm==4.28.1
tweepy==3.6.0
twine==1.12.1
twint==1.1.4.3
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.24
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xkit==0.0.0
yarl==1.2.6
youtube-dl==2018.3.14
zope.interface==4.3.2
